refactor: change file structure; rename sql to query
