+++ src/views/dashboard/locale/en-US.ts
