+++ src/views/dashboard/locale/zh-CN.ts
