+++ src/views/dashboard/modules/favorite.vue
