+++ src/views/dashboard/query/editor.vue
