+++ src/views/dashboard/scripts/py-editor.vue
